% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_visit.R
\name{calc_visit}
\alias{calc_visit}
\title{Calculation of visits around a point.}
\usage{
calc_visit(current_center, sfdf, radius = NULL, maxtime = NULL,
  time_out = NULL, timecol = NULL, units = NULL, proj4string = NULL)
}
\arguments{
\item{current_center}{spatialPoints on which visit calculations should be done}

\item{radius}{radius for visit}

\item{maxtime}{maximum time to consider relocation as successive and add time before first point and after last point proportional to the part of the segment inside the circle.}

\item{time_out}{maximum time individuals are allowed to leave the circle without considering a new visit}

\item{timecol}{name of spdf@data time column}

\item{units}{unit used for maxtime and time_out}

\item{proj4string}{projection string. Default UTM35S.}

\item{spdf}{spatialPointsDataFrame with movement data.}
}
\value{
a named list with data.
}
\description{
Takes a points and calculates the number, total, mean and median time of visits inside a given distance of this point.
}
\examples{
calc_visit(current_center, spdf ,radius = 500, maxtime = 5, time_out = 2, timecol = "dateTime", units = "hour", proj4string = "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs")
}
