% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generate_grid}
\alias{generate_grid}
\title{Generate an empty grid based on total extent of data}
\usage{
generate_grid(spdf, cellsize = 1000, expand = 1.05)
}
\arguments{
\item{spdf}{movement data}

\item{cellsize}{resoluto}

\item{proj4string}{projection string. Default UTM35S.}
}
\value{
a grid
}
\description{
Returns a grid containing all data provided.
}
\examples{
generate_grid(df, cellsize = 1000)
}
