% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{map_residence}
\alias{map_residence}
\title{Calculate number of visit and residence time for movement data}
\usage{
map_residence(df, radius = NULL, time_out = NULL, basename = NULL,
  driver = "GeoTIFF", coord.names = c("x", "y"), timecol = "dateTime",
  grid = NULL,
  proj4string = "+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  cellsize = NULL, progress = T)
}
\arguments{
\item{df}{movement data}

\item{radius}{for finding residence event}

\item{time_out}{maximum time (in s) individuals are allowed to leave the circle without considering a new visit.}

\item{basename}{rasters will be saved as basename_nvisit.tif,basename_totRT.tif,basename_meanRT.tif,basename_medianRT.tif}

\item{driver}{for rasters}

\item{coord.names}{coordinate names in df data.frame. default x,y}

\item{timecol}{name of time column in df data.frame}

\item{grid}{if you want to provide a raster for calculations.}

\item{proj4string}{projection string. Default UTM35S.}

\item{cellsize}{if you dont provide any grid argument, size for the cell used for calculations. Else, will generate a grid with 50 cells for the highest dimension.}

\item{progress}{(boolean) should the function display a progress bar}
}
\value{
a named list of raster object
}
\description{
Creates 4 maps with : number of visit in a given radius, total time spent, mean and median time per visit.
}
\examples{
map_residence(df, radius = 500, maxtime = 2, time_out = 1, basename = "file",
coord.names = c("x","y"), timecol = "dateTime",units ="hour", proj4string =
"+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84
+towgs84=0,0,0",cellsize = 250, progress = T)
}
