% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R, R/mapvisit.R
\docType{package}
\name{mapvisit}
\alias{mapvisit}
\alias{mapvisit}
\alias{mapvisit-package}
\title{Calculate number of visit and residence time for movement data}
\usage{
mapvisit(df, radius = NULL, maxtime = NULL, time_out = NULL,
  basename = NULL, driver = "GeoTIFF", coord.names = c("x", "y"),
  timecol = "dateTime", grid = NULL, units = "min",
  proj4string = "+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  cellsize = NULL, progress = T)
}
\arguments{
\item{df}{movement data}

\item{radius}{for finding residence event}

\item{maxtime}{maximum time to consider relocation as successive and add time before first point and after last point proportional to the part of the segment inside the circle.}

\item{time_out}{maximum time individuals are allowed to leave the circle without considering a new visit}

\item{basename}{rasters will be saved as basename_nvisit.tif,basename_totRT.tif,basename_meanRT.tif,basename_medianRT.tif}

\item{driver}{for rasters}

\item{coord.names}{coordinate names in df data.frame. default x,y}

\item{timecol}{name of time column in df data.frame}

\item{grid}{if you want to provide a raster for calculations.}

\item{units}{unit used for maxtime and time_out}

\item{proj4string}{projection string. Default UTM35S.}

\item{cellsize}{if you dont provide any grid argument, size for the cell used for calculations. Else, will generate a grid with 50 cells for the highest dimension.}

\item{progress}{(boolean) should the function display a progress bar}
}
\value{
a named list of raster object
}
\description{
Creates 4 maps with : number of visit in a given radius, total time spent, mean and median time per visit.

mapvisit helps calculate a map of number of visit and mean, total and median
residence time in a given radius for movement data. It takes into account
maximum time between locs to consider them as successive. It also consider
that the animal can leave and re-enter the circle.
}
\examples{
mapvisit(df, radius = 500, maxtime = 2, time_out = 1, basename = "file", coord.names = c("x","y"), timecol = "dateTime",units ="hour", proj4string = "+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",cellsize = 250, progress = T)
}
