% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_mapvisit}
\alias{run_mapvisit}
\title{Runs map_residence for a list individuals}
\usage{
run_mapvisit(df, listIndiv, folder, grid, basename, IDcol = "id",
  timecol = "dateTime", coord.names = c("x", "y"), ...)
}
\arguments{
\item{spdf}{movement data}

\item{cellsize}{resoluto}

\item{proj4string}{projection string. Default UTM35S.}
}
\value{
a grid
}
\description{
Returns a grid containing all data provided.
}
\examples{
generate_grid(df, cellsize = 1000)
}
